the first version of the code generator
