+++ src/main/java/ru/example/Main.java
@@ -1,0 +1,9 @@
+package ru.example;
+
+import ru.example.other.TestCaseTemplate;
+
+public class Main {
+    public static void main(String[] args) throws Exception {
+        TestCaseTemplate.generateTestCase();
+    }
+}
