+++ src/main/resources/templates/FreeMarkerScreenClassTemplate.ftl
@@ -1,5 +1,5 @@
 package ru.example.screens;
 
 public class ${className} {
-    ${classFilling}
+${classFilling}
 }
