--- src/main/resources/templates/template.ftl
@@ -1,3 +1,0 @@
-PM Name : ${pmName}
-
-Country : ${country}
